feat: Add package manager to npm package.json
This ensures all devs use pnpm and use the same version. The README
explains we must use pnpm (but not the version).
It picks a recent stable version and pins that.
Other devs running other versions will now get an error.

Running a recent pnpm version will add this attribute. So if not added
now, a future commit may add this accidentally or devs will need to
ignore and revert this all the time.

This increases reproducability of builds